Exercicio de fixacao para formatacao de variaveis

+++ src/Exec01.java
@@ -9,7 +9,7 @@ public class Exec01 {
         String product1 = "Computer";
         String product2 = "Office desk";
 
-        int age = 30;
+        byte age = 30;
         int code = 5290;
         char gender = 'F';
